+++ lotus/test/providers/cpu/nn/conv_op_test.cc
@@ -36,7 +36,7 @@ void TestConvOp(const ConvOpAttributes& attributes,
 
   LOTUS_ENFORCE(inputs.size() <= 3, "Our name array is only setup to handle 3 inputs");
   const char* szNames[] = {"X", "W", "B"};
-  for (int i = 0; i < inputs.size(); i++) {
+  for (size_t i = 0; i < inputs.size(); i++) {
     test.AddInput<float>(szNames[i], input_shapes[i], inputs[i]);
   }
   test.AddOutput<float>("Y", expected_output_shape, expected_output);
